requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0

# Database & Vector Store
faiss-cpu>=1.8.0
sqlalchemy>=2.0.25

# LLM & Embeddings
ollama>=0.1.6
langchain>=0.1.6
langchain-community>=0.0.16

# Document Processing
PyMuPDF>=1.23.21
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.2.0

# System Integration
pynput>=1.7.6
pyperclip>=1.8.2
mss>=9.0.1

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
